add new methods to user

+++ bambook_api/app/api/api_bambook/v1/users.rb
@@ -6,6 +6,12 @@ module ApiBambook
           { user: 'Hello' }
         end
 
+        desc 'Return list of users'
+        get do
+          users = User.all
+          present users, with: ApiBambook::Entities::UsersEntity
+        end
+
         desc 'Create a new user.'
         params do
           requires :user, type: Hash do
@@ -42,12 +48,27 @@ module ApiBambook
           end
         end
 
+        desc 'Return a specific user'
+        route_param :id do
+          get do
+            user = User.find(params[:id])
+            present user, with: ApiBambook::Entities::UsersEntity
+          end
+        end
+
         desc 'Delete user'
+        params do
+          optional :Authorization, type: String, documentation: { param_type: 'header' }
+        end
         route_param :id do
           delete do
             user = User.find(params[:id])
-            user.destroy
-            { status: :deleted }
+            if current_user?(user)
+              user.destroy
+              { status: :deleted }
+            else
+              { status: :no_access }
+            end
           end
         end
       end
